backend/requirements.txt
flask
requests
google-generativeai
python-dotenv
readability-lxml
Flask-Cors
gunicorn